requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==5.0.4
django-colorfield==0.11.0
django-emoji==2.2.2
django-markdownx==4.0.7
django-mdeditor==0.1.20
filelock==3.14.0
Markdown==3.6
mistune==3.0.2
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
virtualenv==20.25.3
